Link component, small changes, errors fixed

+++ src/store/productSlice/deleteFromDB/deleteFrofDB.ts
@@ -1,13 +1,19 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import axios from 'axios';
+import {ThunkConfig} from '../../types';
 
-export const deleteFrofDB = createAsyncThunk(
+export const deleteFrofDB = createAsyncThunk<void, number, ThunkConfig<string>>(
 	'deleteProducts/deletefromDB',
-	async (id: number) => {
+	async (id: number, thunkAPI) => {
 		try {
-			await axios.delete(`http://localhost:3000/products/${id}`);
+			if (id) {
+				const response = await axios.delete(`http://localhost:3000/products/${id}`);
+				return response.data;
+			} else {
+				throw new Error();
+			}
 		} catch (e) {
-			console.log(e);
+			return thunkAPI.rejectWithValue('error while deleting from DB');
 		}
 	}
 );
